backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
torch==2.5.0
transformers==4.35.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
